requirements.txt
# Garden-Tiller Python Dependencies

# Core requirements
ansible-core==2.16.14
jinja2>=3.0.0
pyaml>=21.10.1
netaddr>=0.8.0

# Resilience and error handling
pybreaker>=0.6.0

# Hardware interaction
python-ipmi>=0.5.0
redfish>=3.1.0
python-ilorest-library>=3.3.0
git+https://github.com/dell/iDRAC-Redfish-Scripting.git#egg=idrac-redfish-scripting
proliantutils>=3.1.0

# Network tools
dnspython>=2.2.0
pynetbox>=6.6.0

# Reporting
matplotlib>=3.5.0
networkx>=2.7.0

# Structured logging
structlog>=23.1.0
colorama>=0.4.4
rich>=12.0.0

# Utilities
paramiko>=2.10.0
requests>=2.27.0
urllib3>=1.26.0
